Revert d035ac3..e745e32
This rolls back to commit d035ac3bd145f6b28574eec6ebf74b0e98d015d9.

+++ CinemApp/RateViewController.m
@@ -134,7 +134,7 @@ UITapGestureRecognizer *tap;
         
         //Lägger cast i en array
         castArray = [creditsJson objectForKey:@"cast"];
-        NSUInteger castInt;
+        int castInt;
         if([castArray count] > 7)
             castInt = 7;
         else
